+++ library/core/engine/src/main/kotlin/de/lise/fluxflow/engine/workflow/WorkflowServiceImpl.kt
@@ -1,120 +1,60 @@
 package de.lise.fluxflow.engine.workflow
 
-import de.lise.fluxflow.api.event.EventService
-import de.lise.fluxflow.api.workflow.Workflow
-import de.lise.fluxflow.api.workflow.WorkflowIdentifier
-import de.lise.fluxflow.api.workflow.WorkflowNotFoundException
-import de.lise.fluxflow.api.workflow.WorkflowService
+import de.lise.fluxflow.api.ReferredWorkflowObject
+import de.lise.fluxflow.api.continuation.Continuation
+import de.lise.fluxflow.api.workflow.*
 import de.lise.fluxflow.api.workflow.query.WorkflowQuery
-import de.lise.fluxflow.api.workflow.query.filter.WorkflowFilter
-import de.lise.fluxflow.api.workflow.query.sort.WorkflowSort
-import de.lise.fluxflow.engine.event.workflow.WorkflowCreatedEvent
-import de.lise.fluxflow.engine.event.workflow.WorkflowDeletedEvent
-import de.lise.fluxflow.engine.job.JobServiceImpl
-import de.lise.fluxflow.engine.step.StepServiceImpl
-import de.lise.fluxflow.persistence.workflow.WorkflowPersistence
-import de.lise.fluxflow.persistence.workflow.query.toDataQuery
-import de.lise.fluxflow.query.Query
-import de.lise.fluxflow.query.filter.Filter
 import de.lise.fluxflow.query.pagination.Page
 import kotlin.reflect.KClass
 
-
 class WorkflowServiceImpl(
-    private val persistence: WorkflowPersistence,
-    private val eventService: EventService,
-    private val stepService: StepServiceImpl,
-    private val jobService: JobServiceImpl,
-    private val activationService: WorkflowActivationService
+    private val workflowStarterService: WorkflowStarterService,
+    private val workflowQueryService: WorkflowQueryService,
+    private val workflowUpdateService: WorkflowUpdateService,
+    private val workflowRemovalService: WorkflowRemovalService
 ) : WorkflowService {
-    fun <TWorkflowModel> create(
+    override fun <TWorkflowModel, TContinuation> start(
         workflowModel: TWorkflowModel,
+        continuation: Continuation<TContinuation>,
+        originatingObject: ReferredWorkflowObject<*>?,
         forcedId: WorkflowIdentifier?
     ): Workflow<TWorkflowModel> {
-        val data = persistence.create(workflowModel, forcedId)
-        val workflow = activationService.activate<TWorkflowModel>(data)
-        eventService.publish(WorkflowCreatedEvent(workflow))
-
-        return workflow
+        return workflowStarterService.start(
+            workflowModel,
+            continuation,
+            originatingObject,
+            forcedId
+        )
     }
 
     override fun getAll(): List<Workflow<*>> {
-        return persistence.findAll()
-            .map {
-                activationService.activate<Any?>(it)
-            }
+        return workflowQueryService.getAll()
     }
 
     override fun <TWorkflowModel : Any> getAll(workflowType: KClass<TWorkflowModel>): List<Workflow<TWorkflowModel>> {
-        return getAll(
-            workflowType,
-            Query.empty<WorkflowFilter<TWorkflowModel>, WorkflowSort<TWorkflowModel>>()
-                .withFilter(
-                    WorkflowFilter(
-                        id = null,
-                        model = null,
-                        modelType = Filter.ofType(workflowType)
-                    )
-                )
-        ).items
+        return workflowQueryService.getAll(workflowType)
     }
 
     override fun getAll(query: WorkflowQuery<*>): Page<Workflow<*>> {
-        return persistence.findAll(
-            query.toDataQuery()
-        ).map {
-            activationService.activate<Any?>(it)
-        }
+        return workflowQueryService.getAll(query)
     }
 
-    @Suppress("UNCHECKED_CAST")
     override fun <TWorkflowModel : Any> getAll(
         modelType: KClass<TWorkflowModel>,
         query: WorkflowQuery<TWorkflowModel>
     ): Page<Workflow<TWorkflowModel>> {
-        return getAll(
-            (query as WorkflowQuery<*>).mapFilter {
-                (it ?: WorkflowFilter.empty<TWorkflowModel>())
-                    .fallbackModelTypeFilter(
-                        Filter.ofType(modelType)
-                    )
-            }
-        ).map {
-            it as Workflow<TWorkflowModel>
-        }
+        return workflowQueryService.getAll(modelType, query)
     }
 
     override fun <TWorkflowModel> get(identifier: WorkflowIdentifier): Workflow<TWorkflowModel> {
-        return persistence.find(identifier)
-            ?.let {
-                activationService.activate(it)
-            } ?: throw WorkflowNotFoundException(identifier)
+        return workflowQueryService.get(identifier)
+    }
+
+    override fun <TWorkflowModel> saveChanges(workflow: Workflow<TWorkflowModel>): Workflow<TWorkflowModel> {
+        return workflowUpdateService.saveChanges(workflow)
     }
 
     override fun delete(identifierToDelete: WorkflowIdentifier) {
-        val workflowDataToDelete = persistence.find(identifierToDelete)
-            ?: throw WorkflowNotFoundException(identifierToDelete)
-
-        val workflowToDelete = activationService.activate<Any?>(workflowDataToDelete)
-
-        persistence.delete(identifierToDelete)
-        stepService.deleteAllForWorkflow(identifierToDelete)
-        jobService.deleteAllForWorkflow(identifierToDelete)
-
-        eventService.publish(
-            WorkflowDeletedEvent(
-                workflowToDelete
-            )
-        )
-    }
-
-    /**
-     * Deletes the workflow with the given identifier and all of its related steps.
-     * In contrast to the [delete] function, it does not delete jobs nor does it publish a [WorkflowDeletedEvent].
-     * @param identifierToDelete the id of the workflow to delete
-     */
-    internal fun removeSilently(identifierToDelete: WorkflowIdentifier) {
-        persistence.delete(identifierToDelete)
-        stepService.deleteAllForWorkflow(identifierToDelete)
+        return workflowRemovalService.delete(identifierToDelete)
     }
 }