+++ library/core/engine/src/test/kotlin/de/lise/fluxflow/engine/continuation/ContinuationServiceTest.kt
@@ -34,6 +34,7 @@ class ContinuationServiceTest {
             continuationHistoryServiceImpl,
             mock {},
             mock {},
+            mock {},
             mock {}
         )
         val workflowId = mock<WorkflowIdentifier> { }